remove pan responder
I previously had a pan responder on the side menu container in order to capture the ‘tap outside the visible menu area to close it’ functionality. But this was preventing the scroll view child of that container from working properly. So I removed the pan responder and instead added a “clear” view that’s a sibling to the menu. It captures the tap event and closes the menu when tapped.