Side menu (#29)
* wip - working sidebar

* wire up action + reducer

* fix persistent open bug by dispatching action when drawer closes

* cleanup + test

* after rebase, discovered bug with drawer lib when swiping among stories

* chapter action/constant/reducer + select chapter from side menu

* roll our own side menu drawer

* tests

* refresh view pager data source when the currently selected chapter is changed in the side menu

* convert tabs to spaces

* add expanatory comment

* after rebase w/ master

* code review changes

- Move chapter list into its own component
- Clean up view pager data source initialization
- remove unneeded styles

* remove pan responder

I previously had a pan responder on the side menu container in order to capture the ‘tap outside the visible menu area to close it’ functionality. But this was preventing the scroll view child of that container from working properly. So I removed the pan responder and instead added a “clear” view that’s a sibling to the menu. It captures the tap event and closes the menu when tapped.

* code review comments - remove underscores from method names